Convert Float To BigDecimal

+++ src/main/java/com/recurly/v3/requests/AccountAcquisitionCost.java
@@ -9,13 +9,14 @@ import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 import com.recurly.v3.Request;
 import com.recurly.v3.resources.*;
+import java.math.BigDecimal;
 
 public class AccountAcquisitionCost extends Request {
 
   /** The amount of the corresponding currency used to acquire the account. */
   @SerializedName("amount")
   @Expose
-  private Float amount;
+  private BigDecimal amount;
 
   /** 3-letter ISO 4217 currency code. */
   @SerializedName("currency")
@@ -23,12 +24,12 @@ public class AccountAcquisitionCost extends Request {
   private String currency;
 
   /** The amount of the corresponding currency used to acquire the account. */
-  public Float getAmount() {
+  public BigDecimal getAmount() {
     return this.amount;
   }
 
   /** @param amount The amount of the corresponding currency used to acquire the account. */
-  public void setAmount(final Float amount) {
+  public void setAmount(final BigDecimal amount) {
     this.amount = amount;
   }
