+++ src/main/java/com/recurly/v3/requests/Pricing.java
@@ -9,6 +9,7 @@ import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 import com.recurly.v3.Request;
 import com.recurly.v3.resources.*;
+import java.math.BigDecimal;
 
 public class Pricing extends Request {
 
@@ -20,7 +21,7 @@ public class Pricing extends Request {
   /** Unit price */
   @SerializedName("unit_amount")
   @Expose
-  private Float unitAmount;
+  private BigDecimal unitAmount;
 
   /** 3-letter ISO 4217 currency code. */
   public String getCurrency() {
@@ -33,12 +34,12 @@ public class Pricing extends Request {
   }
 
   /** Unit price */
-  public Float getUnitAmount() {
+  public BigDecimal getUnitAmount() {
     return this.unitAmount;
   }
 
   /** @param unitAmount Unit price */
-  public void setUnitAmount(final Float unitAmount) {
+  public void setUnitAmount(final BigDecimal unitAmount) {
     this.unitAmount = unitAmount;
   }
 }
